Adding support to build AzureAD Adapter and also AzureADPreview (#53)
Adding support to build AzureAD and AzureADPreview versions of the adapter.

+++ build/Create-CompatModule.ps1
@@ -2,21 +2,23 @@
 #  Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
 # ------------------------------------------------------------------------------
 [cmdletbinding()]
-param($TargetDirectory = $null, [switch] $noclean)
+param($TargetDirectory = $null, $Module = "AzureAD", [switch] $noclean)
 
 . (join-path $psscriptroot "/common-functions.ps1")
 . (join-path $psscriptroot "../src/CompatibilityAdapter.ps1")
 
-Remove-BuildDirectories
-$mapper = [CompatibilityAdapterBuilder]::new()
+if($noclean -eq $true) {
+    Remove-BuildDirectories
+}
 
-$customizationFiles = Get-CustomizationFiles
+$mapper = [CompatibilityAdapterBuilder]::new($Module)
+
+$customizationFiles = Get-CustomizationFiles -Module $Module
 foreach($file in $customizationFiles){
     $cmds = & $file
     $mapper.AddCustomization($cmds)
 }
-
-$AdditionalFunctions = Get-CustomizationFiles -Directory 'AdditionalFunctions'
+$AdditionalFunctions = Get-CustomizationFiles -Directory 'AdditionalFunctions'  -Module $Module
 foreach($file in $AdditionalFunctions){
     $mapper.AddHelperCommand($file)
 }
